Resolve "Utility tracker fails for very high simulation numbers in verifier"

Closes #108

See merge request ge69hoj/sequential-auction-on-gpu!88

+++ src/verifier/mean_utility_tracker.py
@@ -9,7 +9,7 @@ class UtilityTracker(object):
             1, device=device, dtype=torch.float, requires_grad=requires_grad
         )
         self.cum_sim_count = torch.zeros(
-            1, device=device, dtype=torch.int, requires_grad=requires_grad
+            1, device=device, dtype=torch.long, requires_grad=requires_grad
         )
 
     def add_utility(self, utilities: torch.Tensor):
